Extract email and phone field
